Following D. Smith instructios in order to support xrootd-qs4 on CC-IN2P3 cluster (qms support still need to be added):
"Anyway, the new tag of qserv to install is "u/danielw/lst2013_2", so
you need to get this with:

git clone http://dev.lsstcorp.org/git/LSST/DMS/qserv
cd qserv
git checkout u/danielw/lst2013_2

The new xrootd code for use is xrootd-qs4.tar.gz.  I have put
this in the main download place for qserv, but also I have put it
with all the needed code in:

/sps/lsst/Qserv/src/Tarballs

But there are some details, there is a new qserv-install for
this, since the installation of xrootd has changed.  I have put
this in:

/sps/lsst/Qserv/src/scripts/qserv-install

You will need this to get xrootd to compile, and get installed.
Also the new xrootd requires cmake.  I have the source code
for cmake in the Tarballs dir, but it is probably better to get
cmake installed with yum on the cluster.

Once that is installed, the start_xrootd script needs to change
a bit, xrootd and cmsd are in just $install_dir/bin now, and not
in the x86_64_linux_26_dbg dir, so you have to remove that
from the path.

There is also the fact that the qserv metadata server, the qms
will need to be running before the master can get running,
but that is probably more detail than I want to get into right
now, it would be good to just get the code installed.  I can
startup and fill the qms database tomorrow then. "