Make non-partitioned queries work again. DUMMY_CHUNK now a global const.
Patch addChunk to be cognizant of hasChunks() and reject chunkspecs when
a query is non-chunked. We don't want to bubble up the chunked/non-chunked
information up to the python level. When finalizing queries, if no chunks
have been added, add the dummy chunk.
To be revised after C++ geom lands.
Make MySqlExportMgr always publish the dummy chunk.