Implement reading of packed items from zk/files
open iface for kvinterface placed into facade

Change signature of ConfigMap constructor to attach to a StringMap ref
rather than making a deep copy. This is somewhat less safe in general, but
absolutely fine in our current usage, and clarifies the role of ConfigMap
as a thin veneer rather than a standalone container.

Use config info to use kazoo to populate kvmap and construct facade snapshot