Initial implementation of the new data loading script.
New loading procedure is implemented as a small bunch of python modules:
  configParser.py - parses "new-style" partitioner configuration files
  partConfig.py   - extracts relevant parameters from partitioner config
  dataLoader.py   - the whole procedure for partitioning data, loading
                    it into database and updating CSS

In addition to the modules that can be imported and used independently
there is a new script data-loader.py which accepts a bunch of command
line options and arguments and calls dataLoder.py to do actual data
loading.

See also DM-621 for discussion.