Fix parenthesization in rendered IR tree (DM-737)
This fixes parethesization in rendering of BoolTerm trees.  A
BoolTerm::OpPrecedence enum was added, and the various subclasses
of BoolTerm were assigned relative precedence based on their operators
per the SQL92 standard.  The term list render helper
template function was extended to add parentheses when rendering a
contained term iff the operator of the contained term is of lower
precedence than that of the containing term.

A regression test for the reported test case, and some additional
unit test for parenthesis rendering were also added.