Add workerAdmin module for worker management (DM-1616).
New module workerAdmin defines class WorkerAdmin which implements remote
worker access API based on SSH tunneling (where required). Note that
current API exposes mysql server (it provides a tunnel for mysql
connection) and allows direct execution of commands on remote machine.
In the future this will likely be replaced with special service
(e.g. http-based) and API for interacting with it.

API defined in this module is low-level and exposes things that are
likely to change, we'll need higher-level stable API, but I have no idea
yet what has to be in that API, will extend WorkerAdmin as we discover
what we need.

Unit test for the new module is added as well.