Add support for creating new database on worker nodes.
New module workerMgmt has been added to admin package which is
responsible for multi-node management operations. It allows flexible
selection of a set of workers based on node type and node state.
Currently only one operation is defined - createDb() - which creates
database on all selected workers and optionally grants user access to
this new database. Interface for this method will likely change in the
future if we move from direct mysql interface to something different.