Add createTable() method to WorkerMgmt class.
This new method follows the same pattern as createDb() method added
earlier. It takes the arguments which select a subset of worker nodes,
arguments that define mysql connection options (these are likely to be
replaced in the future) and database/table names. Schema description
must be defined in CSS prior to calling this method.