Add query metadata schema and update configure script to load it.
Query Metadata schema is defined in a new SQL file which is to be loaded
by qserv-configure script together with other schemas. For now we reuse
qservMeta database for that (same schema that is used for seconday
indices), the name of the schema is hardcoded in schema file and the
file itself was generated by Mysql Workbench.