Implement test application for query metadata.
It is implemented as boost unit test, but it cannot be run as a unit
test. To run it one needs mysql server up and running. It needs to
create a new empty database so it needs root priviliges (and is asks for
mysql root password). It loads the schema from QservMetadata.sql file so
it has to run from qserv source top directory (path to SQL file is
hardcoded).