Add QMeta updates to zcar (DM-3163)
Czar is now updating query metadata database when it processes query.
Some pieces are still missing though - we do not have user information
for example, and all queries are SYNC currently, so this will need to be
extended later when we work on those parts. Also missing is worker
assignments to query chunks, I have not managed to figure out yet which
part of code has worker info, to be done later as well (and that is only
relevant to ASYNC queries and only when we need job migration between
czars, so it can wait).

Czar script now has one additional option which can set czar name, by
default czar name is built from host name and czar port number. Czar
configuration file was also extended to include parameters for qmeta
database connection options (currently the same as czar database).