Adding check for table existence in CSS (DM-3163)
Facade::containsTable() method was updated to not only check key
existence in CSS but also to check key value, table is assumed to exist
only if key value is "READY", false is returned for all other values.

UserQuery::submit() now checks that tables exist after registering
registering a query in qmeta database, it table does not exist (e.g.
table was marked as "to be deleted") query is marked as failed. This
needed some updates to UserQuery::join() method as well because of
complicated logic of error/exception handling.

Currently this code does not do much because we are still using CSS
snapshot which is read-only and table existence is checked even before
query is submitted.