Fix worker side sub-chunk reference counting.
The code for tracking sub-chunks in use by a worker was passing
sub-chunk reference counts where sub-chunk IDs are expected
(both are of type int). Because sub-chunk tables are supposed to
be deleted when their reference counts reach 0, the effect is that
sub-chunk 0 is repeatedly dropped, and eventually an assertion gets
triggered when sub-chunk 0 is missing but expected to be present.