Extend KvInterface API to support operations needed by CssAccess
CssAccess needs better control over consistency of the data that it
receives from KvInterface, to support that new method
KvInterface::getMany() was added which is supposed to return a consisten
set of values for multiple keys.

Implement creation of unique keys in KvInterface very much like
zookeeper was doing it by adding unique 10-digit suffix at the end of
the key.