Initial watcher re-implementation based on new CSS (DM-3192).
Core of the functionality is in few classes in admin/watcherLib module,
I tried to make it testable enough, unit test for these are added too.
The watcher is not running yet as a service, will need more testing and
integration with the rest of qserv (data loader initially) and it will
be done in a separate ticket.