Small-scale cleanup for dead Python code.
All Python code will disappear at some point so there will be a larger
cleanup. This cleanup is only to help me to understand how to implement
DROP TABLE better on server side. There is still plenty of Python code
in modules/czar which is not used and is broken.