Remove dependency on MySQLdb module (DM-3949)
Updated wmgr code to use sqlalchemy instead of MySQLdb module.

Same is done for tests/MySqlUdf.py, plus updated that test script for
most recent scisql function names and signatures.

core/examples/obsolete/loader.py still has an import for MySQLdb, we
keep that obsolete script just in case someone wants to look at it but
we do not use it anywhere.

Note that I did not remove mysqlpython from ups table file as we still
depend on MySQLdb as a driver for sqlalchemy. We may switch to some
other driver in the future, then we can replace this dependency.