Force correct library path on OS X
On OS X El Capitan the DYLD_LIBRARY_PATH environment variable
will be stripped from the environment as part of Apple's
System Integrity Protection. We therefore have to copy
it from the "safe" LSST_LIBRARY_PATH variable and ensure
that the tests are run with this path set explicitly as
part of the invocation (since the subprocess may also
lose the path).

Furthermore, Qserv libraries are not built with @rpath
set and test binaries are not set up to use a relative
rpath, we therefore force the "build" directory into
the library search path.