Implement unique query-id generation (DM-5314)
Query ID is now generated by QMeta when registering new query and it is
globally unique. This query ID is not used for the message table name
which is still generated from the same "suer query ID", the reason is
that message table should be created before the rest of the system kicks
in and message table must exist even if we don't register (or fail to
register) query in QMeta.

As a small optimization czar now avoids creation of the result table
when it is not used (for some non-SELECT queries). Proxy Lua logic is
updated to handle empty result table name.