-install_name is understood by OS X clang compiler
There is no need to use -Wl, to pass "-install_name" to the linker.
Clang on OS X already knows how to do this so the command line looks
cleaner without it. This change also prepends @rpath to the install
path to make it easier for other applications to specify where
the libraries might be without relying on DYLD_LIBRARY_PATH.