update testQueryAnaGeneral for subchunk generation in worker
the subchunk queries are no longer being generated in the czar,
this test was assuming they were. This changes the verifcation to
look for the correct number of chunk queries, and allows for %SS%
to be left in the query string, since it would be replaced later
by the worker.

TODO I think it would be good to add a verificaiton that the
subchunks to query are present in whatever becomes the
TaskMsg.Fragment.subchunk.id vector, since the new mechanism
depends on those numbers being there for the worker to generate
the queries for those subchunks.