remove QueryMapping::insertSubChunkEntry
Also add docstrings to other funcs in that class.

We no longer want QueryMapping, in the czar, to replace subchunk
entries in the query strings. Instead it will be handled in the
worker. Removing this function prevents any accidental use that
would cause the subchunk entry to be marked for replacement in
the Czar.