Fix integration test worker crash for MariaDB 10.2
Under MariaDB 10.2.14, the MYSQL_FIELD def_length field is sometimes
coming back garbage (bignum), causing a std::string ctor below to
blow up.

MySQL docs say this field and its partner are not set unless this
MYSQL_FIELD comes back from a call to mysql_list_fields(), which is
not the case in our call chain. Perhaps this field happened to be
init'd or harmless in previous versions of MariaDB, but it seems no
more...

Removed the default-value machinery from sql::Schema, as it was
thus never correct and was not being used.  Deprecated corresponding
fields in the ColumnSchema protobuf message.