simplify fetching valid column names
the `if` is redunant; vf->getType(); and the switch guarantees
that at that location the ValueFactor is a column ref.

call getColumnRef on the ValueFactor instead of the ValueExpr
because the ValueExpr just calls the ValueFactor and calling
the ValueFactor is explicit and so in this case is more clear.