add antlr4 support for nested expression in WHERE
elimiates the requirement (at least temporarily) that the
a4-generated IR be exactly the same as the old/v2 antlr
generated IR. There is still a requirement that both IRs
create exactly the same query when rendered.
This is done because the old antlr parser added extra terms to
the IR that then went through a 'reduction' step to remove
extraneous terms.
The alternative to removing the 'equals' requirement would be to
reduce both IRs before comparing them. I'm not sure that's even
the right way to go at this point.