add support for star in eg. 'SELECT myTable.* ...`
includes an ASSERT in QuerySpecificationAdapter that at least a select
statement and a from clause were created.

also tweaks the definition of REAL_LITERAL grammer element a little, so
that the DEC_DIGIT before and after is optional, but at least one
DEC_DIGIT next to a decimal point is required for it to be a
REAL_LITERAL. That way a decimal without a digit can be treated as a
decimal.