make parser exceptions user friendly & log details.
in parser asserts an exception is thrown and that message is displayed
to the user in the sql prompt, so it should be something meaningful to
the user.
This makes the message in the exception something that will be
informative for the user.
The information that will be useful for debugging gets sent to the log
(in this case that's the czar log)