verify tokens are not silently ignored
adds a required function to Adapter called checkContext where subclasses
can/should check the context to be sure that non-context members are
either expected or null (or throw) so that unexpected tokens do not get
silently ignored. Members that return a context subclass typically do
not need to be checked because there will be a child adpater to handle
that context that will get called by the parse tree walker.