update unit tests now that '%' parses
A query test case with '%' was listed in the failing queries in
testCControl. Now that we support that operator we remove it from that
list.

The query was in that list because it appears in a test in qproc but was
commented out as a FIXME test case. Now that test can expect the parse to
succeed, but now it fails in analysis. The test case has been updated to
expect that analysis error.