deprioritize tokens in lexer
ZERO_DECIMAL, ONE_DECIMAL, and TWO_DECIMAL are used to allow binary
tokens '0' and '1' in the grammar, and also '2' in some cases. But their
priority was higher than decimal literal (0 through 9) and causing
errors in the case of -1 and -2 (but not 0, 1, or 2 for reasons that are
not really clear to me). I'd rather binary tokens parse as decimals than
have special handling for decimals that might be binary tokens so I made
these tokens lower priority than DECIMAL in the lexer.