make unit test with IR build helpers
The test containes factory fucntions with the name of the target that
return the needed type (usually a shared_ptr to a class intance of the
named function), e.g. SelectStmt(...) returns a
shared_ptr<query::SelectStmt>(...).

In cases where an object can own a vector of child objects, variadic
templates were used so that any number of objects may be passed directly
to the parent object without fancy annotations to e.g. initialize a
vector, which became untolerably complicated when the vector had to be
owned by a shared_ptr.

The final reuslt of all this is to make creating the _expeceted_ IR
hierarchy as clear and easy to read as possible.