modify logged IR for test factory
This adds factory functions for IR in ccontrol/testAntlr4GeneratedIR
that allow query IR to be made with code that is as short as possible
but is still explicit and correct.

Then, the debug output of query IR outputs in the same format as the
query IR factory functions, so that an instantiated IR hierarchy should
be able to be printed, copied, pasted into a source code file, and
rebuilt into that same IR object. The inspiration for this comes from
the python 'repr` function.

Tests in ccontrol/testCControl.cc that used to use antlr v2 parser code
to crate IR that was "known good" will be replaced with tests in
ccontrol/testAntlr4GeneratedIR.cc, which use the new factory fucntion
method to generate known good IR.