shorten some of the really long select lists
as observed in code review, some of the select lists are really long,
and take too many lines of code to create. I think once there's more
than 2 or 3 columns to select in the select list it really doesnt matter
if there's 2 or 200. So the ones that had more lines than I could view
on a screen I truncated to 3, sometimes 4 colums to select.