use an exception in UserQuery
for InfileMerger, instead of returnung bool + string output variable,
I decided not to change the it to use an exception becuase the api of that
class for the most part follows the pattern of returning bool + output
variable string. I did this in part for consistency and in part because
it was not obvious what exception to return - the InfileMergerError
exception that is an util::error::Exception uses a (util-level) error
code that does not align nicely with this kind of internal exception.
And so, instead of inventing a new exception pattern for this class I
decided to be consistent with its existing api.