put resultQuery in QInfo & build in factory
Some of the query creation that the Czar was doing was specific to one
or two kinds of queries. Moving some of these things into the factory
for that kind of query allowed some tuning of query construction.

There's a chicken-and-egg problem creating the query information in
query metadata and creating the results table, and creating the result
query that is stored in metadata, and also contains the query id in the
result table name. UserQuerySelect could probably use some refactoring
in this area.