remove Constraint, make everything QsRestrictor
Constraint & QsRestrictor had the same parameters (a string called name
and a vector of string called parameters), and in
QuerySession::getConstraints a vector of QsRestrictor was converted to a
vector of Constraint, and that was used in processing in SecondaryIndex.
I searched around and asked but I couldn't discover a good reason to
have the 2 classes. Our best guess is that Constraint was part of the
early prototype code for Qserv and has just been left there and adpated
around.

The main motivating factor for removing the redundancy is so that
QsRestrictor can be refactored for different type of restrictors;
sublcasses can have meaningfully named variables instead of a vector of
string with zero built in semantic definition.