allow parser to create user query with config
Adds a query config class, essentially a duplicate of the private query
factory _impl. ..if this pattern seems to be the correct choice we could
get rid of UserQueryFactory::Impl and keep a local UserQueryConfig. The
latter was created just to implement proper coupling between classes
(instead of making the private Impl class public somehow). The
UserQueryConfig could be kept by the UserQueryFactory and passed into
the parser.