remove InfileMergerConfig
The variables it contained were duplicated in various locations and its
member variables were populated at different locations, making it
difficult to know what it represented and when it should be out of
scope - no longer needed. Instead of packing a set of somewhat unrelated
variables into the one container it is simpler to pass the variables
from where they originate to the class(es) that use them.