replace ColumnRef copy ctor with clone function
The copy constructor was broken, introduced when ColumnRef was changed
to contain a TableRef instead of using its own strings for db and table,
becasue the copy constructor was only doing a shallow copy of the
TableRef pointer. As a result, changing the copy's table settings later
could have unintended consequences.

This disallows the copy constructor and replaces it with the
qserv-standard 'clone' (deep copy). The other qserv-standard
'copySyntax' (shallow copy) can be implemented if/when needed, as can
the copy constructor.