fix issue where oneTable was treated as paritioned
self.skipPart and self.oneTable were not being tested when setting the
value of self.partitioned, the way the should have been according to
the comment just above the changed line of code.

as a result, qserv-specific columns for chunking were still being added
to 'oneTable' tables (where the data should be loaded without
partitioning), which was causing issues in the integration tests where
the mysql tables had the chunk columns when they should not.