remove old init script, rename current init script
The current init script replication.sql init script was renamed with the
migration script naming pattern.

The older None-to-0 init script that was removed was added for lite, in 7453e20.
It had a bug (the wrong table name for schema version), and it's not really
useful anymore, because any new database can be initialized to the later version
using the more current init script.