create and load table using same connection
Previously each step used a different connection, and
different credentials (one was admin, the other was
user creds). There was a race condition where the db
was created & the function returned, but when the load
table function was called, the server said the db did
not exist yet. Using one connection (and the same user)
seems to serialize access at the db server better.