Switch to using UUID for worker identities in the integration test
Eliminated the string-based 'vnid' option that was used by the entrypoints
for initial configuration of the worker identifiers in the databases.
The new version uses the UUID-based generator for each new instance of
of the database.

The only remaining parameter that's left in this context is 'vnid-config'
that allows all 3 options available in the xrootd/cmsd configurations:
using a static string, reading a value of the vnid from a file, or
using a special plugin.

The Docker compose configuration was also adjusted accordingly.