Implemented the worker redirector (registration) server
The service is implemented as the HTTP REST server responding
to the worker registration ("login"), management (removing workers)
and information (obtaining a collection of registered work services).
No siginificant worker validation will be made by the server. Worker info
will be passed to the Master Replicaton Controller for further validation
and configuring the the Replication/Ingest system.

In the current implementation the service runs as a separate application
process.