mimic dependencies of lsst.log in lsst.utils
lsst.log depends on a couple functions in lsst.utils, and
lsst.utils depends on numpy, but not for functions used by
lsst.log. To reduce the run container size we remove numpy
and "mimic" the functions in lsst.utils, by copying them into
a duplicate module tree and installing into the qserv run image.