use git tag if newer than dockerfile changes
When creating lite qserv image tags, the description of the sha
that has the most recent changes to the relevant dockerfile(s)
is used to tag the image.
This can be confusing if the most recent changes are from a
previous release.
To make it more clear, if the most recent git tag (qserv release)
is newer then the most recent dockefile changes then use a
description of the SHA of the release tag instead.