Dynamic reconfiguration of the worker connections in the messenger
Two important changes in this commit. First of all, in the new implementation,
worker conection parameters are obtained before attempting to establish
a connection. The older implementation was doing it just once in a lifetime
of the connector and caching the parameters at a time when the connector
was being initialized. The new implementation allows automatically tracking
changes in the connection parameters and react on to the changes accordingly
(reconnect to the new host/port). The use case is important because it allow
worker services to move dynamically from one host (and/or port) to another one.

The second improvement allows adding new connectors dynamically as new workers
are getting registered in the Configuration. This is the typical scenario when
Qserv is scaling up (new workers are added w/o restarting the running services).
Note that, removing workers has not been implemented yet. This will be done later.