Migrated worker services not to use worker registry for configuration parameters
In the older implementation, worker services were expecting configuration
parameters to be stored in the shared persistent configuration. Changes made
in this commit decouple workers from the database. Workers are now completelly
configured by values passed via the corresponding command-line parameters (or
their default values of the class ConfigurationSchema). The Replication system's
configuraton database is no longer used to check if the worker identifier is known
in a scope of the given Qserv instance. The last change is needed to allow new
workers to self-register themselves in the cluster.

Similar changes has been also made to the relevant methods of the DatabaseService
implementation because they're used by the workers.

Note that workers are passing their configuration parameters (port numbers,
data folder locations) to the Redirector service so that countrolles would be
able to learn this info from the Redirector if needed. The Master Controller
will (if allowed) register the new workers in the configuration database.