Optimized implementation of the contribution ingest algorithm
The new version allocates the CSV parser for a duration of time needed
to parse a contribution. After that the parser gets immediately destroyed.
This prevents steady memory growth (16 MB per each contribution) of
the worker process that was seen in the previous version of the code.
The memory consumption was caused by the internal buffer allocated
by the parser.
Another (though, the minor) problem (that still needs to be addressed)
is that the ASYNC contribution processors are created by the worker's
REST service, and they are never cleared pon successful completion
of the requests. This was done on purpose to allow ingest workflow
track a status of the contributions.