Added support for copying the query text
Refactored tables on the query monitoring page to allow copying queries
to the clipboard (a fallback solution is offered for cases
when the clipboard is not writeable). Separated query formatting controls
from browsing the table rows in order to allow additional operations
over queries (the first one was mentoned earlier).