Extended and refactored implementation of the Qserv query monitor
The REST service that's used by the Qserv Web Dashboard for monitoring
and searching queris was extended to allow searching for the given
search pattern in the query text using the FULLTEXT index. Both services
also got an option allowing to return extended info on each (past) query,
such as a collection of messages and query statistics recorded in the Czar's
database during query processing, as well as the number of chunks, rows
and bytes in the result sets of the completed queries.
Anoher REST service that was ment to report info on a specific
query (given its unique identifier) has been implemented.
Implementations of both services have been refactored to prevemt code
duplication and to use the advanved MySQL API.