Extended ObjectIndex to support creating/accessing named instances
The return type of the factory method ObjectIndex::instance was changed
from the reference type to the pointer type to allow caching pointers
to the index instances. This is needed to avoid the additional performance
degradation due to the synchronized implementation of the factory
method in the new code. The caching should be fine since the pointers
are immutable.