Streamlined the configuration of the Docker compose-based setup of Qserv
In the new version of the file, locations of the configuration files
for the LSST Logger are passed explicitly to each entry point. The second
improvement was to set up separate config file (LSST Logger) for each
entry point to allow each service have its own configuration.
The third improvement was to add the missing option "--debug"
when starting the Replication Controller. W/o this option the Controller
ignores the external configuration file for the LSST Logger.