Updated the integration test data and config to allow testing RefMatch
Changed test case03 to eliminate the default parameter "default-overlap"
in the partitioned config with just "overlap". The parameter isn't used
by the partitioning tools. Set the overlap to 0.001 which is sufficient
for matching objects when partitioning the RefMatch table RefDeepSrcMatch.

Changed the overlap to 0.001 in the test case04 in sync with case03
to allow haivng both within the same database family.

Converted the table RefDeepSrcMatch to be the partitioned RefMatch
table (before the change, the table had the fully-replicated type).
Extended the data file of the table with 4 columns carrying spatial
coordinates of the matched objects. These columns are needed to
partition the table data using the tool sph-partition-matches.

Added support in the integration testing tools for partitioning
and loading the RefMatch tables.