Extended Configuration service of the Replication/Ingest system
Added a method for changing the replication level of the database families.
Added a general configuration parameter that enforces the upper limit
for the replication level (of any family) by some number. The default limit
is set to 4. Note that the effective (at run-time) limit will be the lesser
of that "hard" limit, the desired replication level specified for the given
family, and the number of workers that are available for the replication.
This safeguard was implemented to prevent common mistakes that could
be made by the Qserv data administrator which could result in the "explosive"
use of the disk storage.
Updated the test config set and the unit test accordingly.