Minor refactory in a hierachy of the Replication job classes
Pushed two highly specialized methods from class Job to the
corresponding subclasses which are the only users of the methods.
Keeping those within the base class Job was problematic for a few
reasons: 1) methods had no other users but the ones that provided
the new homes for the methods, 2) keeping the methods in the base
class was "exploding" (for no practical reason) the compile-time
dependency graph of any modules that were including replica/Job.h.
That would possibly prolong the compilation time of the affected
modules. 3) these extended dependencies also included XRootD/SSI
headers that were breaking compilation of the new unit test (to be
added in the next commit) due to the indirect dependency onto these
headers (though, this might be also a problem with CMake definition
of the unit tests that may need to be corrected).