Added a resource manager for limiting concurrency of ingest requests
There are two flavors of the resource managers. The transient
manager is represented by the class IngestResourceMgrT. It provides
an in-memory dictionary of limits for databases. The limits could
be changed at run-time by the unit test to allow testing
the limits-awared request scheduler (the manager). And there is
the persistent version of the manager which is represented by the class
IngestResourceMgrP. It's meant to be used in production to pull (in practice
- to track) the limits (as they're configured by the ingest workflows) from
the Replication system's database.

Also added a unit test for the transient version of the resource manager.