Extended the ingest config REST service to support request concurrency
Also did a minor refactoring in comment in the documentation of
the command-line parameters of the application class. The previous
version was screwed by clang-format.

The GET version of the service now requires the name of a database to
be passed in the query string rather than in the request's body.
The previous implementation was incorrect.