Extended transient state and the REST API to support auto retries
Note that, at this stage parameters related to the extended
model of the contribution requests are not preserved
in the persistent state of the Replication/Ingest system. Neither
the system is capable of making actual retries as hinted by the extended
model of the requests. This will be implemented in some later commit.