Extended the index retrieval job to report index status in a specialized object
The previous version of the class SqlGetIndexesJob didn't have a convenient
representation for the statuses of indexes found at worker tables. The new
version adds an adaptor that translates the standard result object
of the base class SqlJob into a special representation. The new
representation also has the JSON interface to simplify using the result
by the REST services, for storing it in the database, or caching in memory.