Refactored processing of the "director" index data
The data received from the worker services are now written directrly
into the temporary files by the Controller's requests rather than
by the jobs. This solves three problems that existed in the
previous implementation. Firstly, the new algorithms avoids copying
the data twice in memory. Secondly, the data aren't cached in memory
for the lifespan of the requests, which would reduce the memory
usage by the Controllers. And finally, and most importantly, the new
interface between the request class DirectorIndexRequest and
the job class DirectorIndexJob allows for more sophisticated
implementation of the request class that would support large
results. The current BOOST/ASIO + Protobuf-based protocol limits
the amount of data sent from workers to the controllers by 64 MB.
This limit may be way to low for so,e chunk tables that may have
many million rows. For example, if the primary key (an "objectId")
of the "director" table is based on the 64-bit integer type, and
the columns "chunkId" and "subChunkId" are both based on the 32-bit
integer type, then the total amount of data per row would be about
8 + 4 + 4 = 16 bytes in the binary representation. This would impose
a limit of 4 millon rows of the index data to be harvested by
the worker. In the text representation this limit could be even lower.
The problem will be solved in the next commit in which the request
class will be pulling rows from the chunk tables using the SQL
windowing syntax 'ORDER BY <primary-key> LIMIT <rows>, <offset>'
until the whole table gets scanned. The results will be recorded
by the request into the same temporary file opened and managed by
the request in the temporary staging area. Once all rows will be
read the job that initiated such request would be given a location
o fthe file. The job will initiate loading the content of teh file
into the destination (the "director" index table).