Initiating garbage collection of the intermediate worker requests
The first change in this commit was extending the base design of
the Controller requests to allow disposing intermediate worker
requests. This improvement covers an important scenario when
the same Controller request is making more than one call to a worker
to retreive large quantities of data in a series of chunks (AKA
the "streaming" model). This model is used when the amount of data
to be retrieved from the worker exceed the "hard" limit of the Protobuf
protocol. That is normally done to expedite the garbage collection of
the worker requests and prevent excessive memory build up (or keeping
other resources) at the worker.
The second change was to improve the new version of the "director" index
retreival request to dispose the intermediate worker requests.