Reintroduced the original version of util::Mutex and util::Lock
The new version of the classes is incompatible with expecation of
the Replication & Ingest system's code. In this change, the original
versions of both classes have been reintroduced as replica::Mutex
and replica::Lock. The dependent classes of the module 'replica' have
been migrated to use the newly (re-)introduced classes.