Refactored task monitoring reports
In the previous implementation, task progression timestamps were
reported into the JSON object as strings. This was making a few
problems for displaying the timestamps on the corresponding page
of the Qserv Web Dashboard. In particular, the default values of
the timestamps had to be displayed as the UNIX Epoch times which
resulted in the "noisy" UI. Analyzing and displaying differences
between the timestamps was another problem. The new implementation
adresses this issue by migrating the timestamp representation to
to the numeric format (the number of milliseconds since teh UNIX Epoch).