Add support for the file-based result delivery to worker's CMSD config
Added the corresponding configuration option to the entry-point CLI
for ingecting a value of the result folder into the config file of
of the worker CMSD. The configuration file has been extended as well
to allow serving files from workers via the XROOTD file-based protocol.
The SSI provider class has been extened to recognise result files
at the local filesystems of workers as valid XROOTD resources.