Major refactoring and code cleanup in the worker query processing classes
Split wbase::SendChannelShared into the base class and an implementation

Did the minor refactoring in naming and using XROOTD/SSI streaming channels
These change were required to prepare ground for introducing other
"ChannelShared"-alike implementations of the result processing and delivery
mechanisms.

The refactoring aims at correcting dependencies in a code that creates
and sets up worker tasks, so that task initialization was happening
completelly within the implementation file of wbase/Task.cc rather
than being spread between wbase/Task.cc and wcontrol/Foreman.cc.
Altogether this is meant to improve the observability of the code and
make it easier to maintain.

Also, some minor code cleanup and removal of unused header includes
and forward declarations was made. Thould reduce the compilation
time of the relevant modules.

Got rid of unused members and methods

Classes wbase::ChannelShared (used to be wbase::SendChannelShared)
and wbase::TransmitData had an obsolete mechanism for setting up
result schema in the response messages to be sent to Czar. The current
implementation no longer uses that API. The obsilete API was removed
to avoid confusions as it wasn't obvious where the actual schema
settings were made.

Got rid of std:: in the iomplementation files with "using std".

Guaranteed and enforced synchronization in private methods that require
a lock to be held before calling the methods. In the new code, a cons
reference to a lock is passed around contexts (methods) where such lock
is required. This compiler-enforced technique has proven to work for large
code bases where it would be hard to track a state of the lock by other methods.
The only other alternative method would be to do the run-time inspection of
the lock at the entrance of a method.